+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 2022-03-18
+- Add Dropbox OAuth PKCE example app
 - Released version 4.0.0 of the SDK
     - Deprecates the `slot` property
 - Fixed https://github.com/stripe/stripe-apps/issues/244